recv/send success

+++ crates/cvitek-nic/src/cvitek_defs.rs
@@ -79,7 +79,7 @@ pub const MAC_ENABLE_TX: u32 = 1 << 3; /* Transmitter Enable */
 pub const MAC_ENABLE_RX: u32 = 1 << 2; /* Receiver Enable */
 
 /* Received Poll Demand */
-pub const DMA_XMT_POLL_DEMAND: u32 = 0x00001004;
+pub const DMA_XMT_POLL_DEMAND: usize = 0x00001004;
 
 /* Received Poll Demand */
 pub const DMA_RCV_POLL_DEMAND: u32 = 0x00001008;
